[Note: new sql update: 2642_loot_template.sql] Implement storing and selection loot base at item grouping. DROP able `skinning_loot_template_alternative` not used any more. Patch provided by Alan Oaten. Fixed problem with large item ids in .lookupitem command. Patch provided by SinnerG. Fixed category spells cooldown remembering at spell cast. Not remembered before. Implement option in mangosd to disable/enable beep at mangosd start for protect cat's rights at sleep ;)
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2642 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Level3.cpp
@@ -1332,7 +1332,7 @@ bool ChatHandler::HandleLookupItemCommand(const char* args)
     do
     {
         Field *fields = result->Fetch();
-        uint16 id = fields[0].GetUInt16();
+        uint32 id = fields[0].GetUInt32();
         std::string name = fields[1].GetCppString();
         PSendSysMessage("%d - %s",id,name.c_str());
     } while (result->NextRow());
